[UPD] In theory (no file to test this), It is possible to extract more than 255 analog channels from a C3D file. [UPD] Scaling factors for analog channels are now updated from acquisition' data. The btk::C3DFileIO::m_AnalogUniversalScale is always set to 1 in a C3D exportation. Need to check if this setting is always correct, especially for analog channel with a high gain.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@137 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/IO/btkC3DFileIO.h
@@ -157,7 +157,7 @@ namespace btk
         this->m_Bfs->Write(residualAndMask);
 
       };
-      virtual void WriteAnalog(double v) {this->m_Bfs->Write(static_cast<int16_t>(static_cast<float>(v)));};
+      virtual void WriteAnalog(double v) {this->m_Bfs->Write(static_cast<int16_t>(v));};
 
     };
     class FloatFormat : public Format
